refactor: refactor index components and set alias path
