+++ layout/default.js
@@ -1,28 +1,34 @@
 import React from "react"
-import { PageContainer } from "../components/global"
-import NavBar from "../components/_layout/navbar"
+import { PageContainer } from "components/global"
+import { motion } from "framer-motion"
+import NavBar from "components/layout/navbar"
 import styled from "styled-components"
 
-const PageWrapper = styled.div`
+const PageWrapper = styled(motion.div)`
   display: block;
   overflow: auto;
 `
 
+const PageAnimationProps = {
+  transition: {
+    type: "spring",
+    duration: 1.2,
+  },
+  initial: {
+    y: -32,
+    opacity: 0,
+  },
+  animate: {
+    y: 0,
+    opacity: 1,
+  },
+}
+
 export default function Layout(props) {
   return (
-    <PageWrapper>
-      <NavBar
-        transition={{ type: "spring", duration: 1.2 }}
-        initial={{ y: -32, opacity: 0 }}
-        animate={{ y: 0, opacity: 1 }}
-      />
-      <PageContainer
-        transition={{ type: "spring", duration: 1.2 }}
-        initial={{ y: -32, opacity: 0 }}
-        animate={{ y: 0, opacity: 1 }}
-      >
-        {props.children}
-      </PageContainer>
+    <PageWrapper {...PageAnimationProps}>
+      <NavBar />
+      <PageContainer>{props.children}</PageContainer>
     </PageWrapper>
   )
 }
